3.0.2 Update
Isolate the bath input in chi2_fitgf. It is not a good idea to modify
input parameters in python. The modified bath will be returned by the
function. The original bath is retained. Breaks compatibility with
EDIpy1.Isolate the bath input in chi2_fitgf. It is not a good idea to
modify input parameters in python. The modified bath will be returned by
the function. The original bath is retained. Breaks compatibility with
EDIpy1.0