3.0.7 Update
Updated the handling of the Uloc, Ust, Jh, Jx, Jp global variables.
In EDIpack2.0 > 3.0.4 a new feature has been introduced whereby the
user can pass the interaction operators in a file. As such, the
above input variables lose relevance once the file is read. Instead
the code stores the interaction coefficients in internal arrays which
are of dimension [Norb] for Uloc and [Norb, Norb] for everything else.

To preserve backcompatibility, the handling of these global variables is
as follows:

-if EDIpack <= 3.0.4 or the internal matrices are not initialized, the
global variables point to the input variables. The dimensions are [NORB]
for Uloc and scalars for everything else

-if EDIpack > 3.0.4 and the internal matrices are initialized, the
global variables modify the matrices. The dimensions are as previously
described, but the user can still pass scalars (or shorter arrays in
the case of Uloc). If a scalar is passed, the matrices will be
set with all relevant elements equal to the scalar.
If a matrix is passed for Ust, Jh, Jx, Jp, the dimension must be
[Norb,Norb].